fixed a bunch of stuff. all tests work

+++ app/users.py
@@ -2,7 +2,8 @@
 
 import os
 from flask import Flask, request, jsonify, abort, make_response
-from app import app, db, models, bcrypt
+from sqlalchemy.orm import subqueryload, contains_eager
+from app import app, db, models, bcrypt, session
 from cors import crossdomain
 
 @app.route('/deku/api/users', methods=['GET', 'POST'])
@@ -15,18 +16,35 @@ def users():
         firstName = request.form.get('firstName')
         lastName = request.form.get('lastName')
         email = request.form.get('email')
-        password = request.form.get('password'),
-        university = request.form.get('university');
+        password = request.form.get('password')            
+        university = request.form.get('university')
         if (firstName and lastName and email and password and university):
+            print password
+            print firstName
+            print password
             pw_hash = bcrypt.generate_password_hash(password)
             user = models.User(firstName = firstName,
                                lastName = lastName,
                                email = email,
                                password = pw_hash,
                                university = university)
+            profile = models.Profile()
+            year = request.form.get('year')
+            major = request.form.get('major')
+            classes = request.form.get('classes')
+            biography = request.form.get('biography')
+            if (year):
+                profile.year = year
+            if (major):
+                profile.major = major
+            if (classes):
+                profile.classes = classes
+            if (biography):
+                profile.biography = biography
+            user.profile = profile
             db.session.add(user)
             db.session.commit()
-            return make_response(jsonify(user = user.serialize), 201)
+            return make_response(jsonify(user = user.serialize, profile = user.profile.serialize), 201)
         else:
             abort(400)
     else:
@@ -42,10 +60,26 @@ def user_by_id(user_id):
         else:
             abort(404)
     elif request.method == 'PUT':
-        user = models.User.query.get(int(user_id))
+        print '********----modinfo----'
+        if u'id' not in session:
+            return make_response("not logged in",401)
+        if session[u'id'] != user_id:
+            return make_response("You cannot modify someone else's information",401)
+        
+        result = db.session.query(models.User, models.Profile).filter(models.User.id== user_id).filter(models.User.id==models.Profile.user_id).first()
+        user =  result.User
+        profile = result.Profile
+        #q = models.User.query.filter(models.User.id == user_id).join(models.User.profile)
+        print 'aaaa-----------'
         firstName = request.form.get('firstName')
         lastName = request.form.get('lastName')
         email = request.form.get('email')
+        password = request.form.get('password')
+        university = request.form.get('univ')
+        year = request.form.get('year')
+        major = request.form.get('major')
+        classes = request.form.get('classes')
+        biography = request.form.get('biography')
         if (user):
             if (firstName):
                 user.firstName = firstName
@@ -53,14 +87,31 @@ def user_by_id(user_id):
                 user.lastName = lastName
             if (email):
                 user.email = email
+            if (password):
+                user.password = bcrypt.generate_password_hash(password)
+            if (university):
+                user.university = university
+            if (year):
+                profile.year = year
+            if (major):
+                profile.major = major
+            if (classes):
+                profile.classes = classes
+            if (biography):
+                profile.biography = biography
             db.session.commit()
-            return jsonify(user = user.serialize)
+            return make_response(jsonify(user = user.serialize, profile = profile.serialize),200)
         else:
             abort(404)
-
+            
+            
     elif request.method == 'DELETE':
+        if u'id' not in session:
+            return make_response("not logged in", 401)
         user = models.User.query.get(int(user_id))
         if user is not None:
+            if session[u'id'] != user.id:
+                return make_response("You cannot delete someone elses account",401)
             db.session.delete(user)
             db.session.commit()
             return make_response(("User deleted.", 200))
@@ -80,12 +131,25 @@ def user_authentication():
             if (user):
                 correct_pw = bcrypt.check_password_hash(user.password, password)
                 if (correct_pw):
-                    return jsonify(user = user.serialize)
+                    session[u'id'] = user.id
+                    return make_response(jsonify(user = user.serialize),200)
                 else:
-                    return abort(401)
+                    session.clear()
+                    return make_response("invalid",401)
             else:
-                return abort(404)
+                session.clear()
+                return make_response("user not found",404)
         else:
+            session.clear()
             return make_response(("Email or password missing.", 401))
     else:
+        session.clear()
         pass
+
+
+@app.route('/deku/api/users/logged_in')
+def check_logged_in():
+    if u'id' in session:
+        return make_response('logged in',200)
+    else:
+        return make_response('not logged in',401)
